% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc.R
\name{cvc}
\alias{cvc}
\title{CVC}
\usage{
cvc(err_mat, B = 200, verbose = F)
}
\arguments{
\item{err_mat}{matrix of \code{n} rows and \code{K} models, each entry containing the loss of a particular entry}

\item{B}{number of bootstrap trials}

\item{verbose}{boolean}
}
\value{
vector of p-values, where larger p-values denote more evidence to support the null hypothesis
of the specific model is the best candidate
}
\description{
CVC
}
