% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc.R
\name{cvc_sbm}
\alias{cvc_sbm}
\title{CVC for SBM (with cross validation)}
\usage{
cvc_sbm(err_mat_list, trials, alpha, verbose = T, ncores = NA)
}
\arguments{
\item{err_mat_list}{error matrix list outputted by \code{edge_cv_sbm$err_mat_list}}

\item{trials}{number of trials}

\item{alpha}{significant level}

\item{verbose}{boolean}

\item{ncores}{integer or \code{NA}}
}
\value{
list
}
\description{
CVC for SBM (with cross validation)
}
