% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc.R
\name{cvc_sbm_sample_split}
\alias{cvc_sbm_sample_split}
\title{CVC for SBM (with sample splitting)}
\usage{
cvc_sbm_sample_split(err_mat, trials, alpha, verbose = T, ncores = NA)
}
\arguments{
\item{err_mat}{error matrix outputted by \code{edge_cv_sbm_sample_split$err_mat}}

\item{trials}{number of trials}

\item{alpha}{significant level}

\item{verbose}{boolean}

\item{ncores}{integer or \code{NA}}
}
\value{
list
}
\description{
CVC for SBM (with sample splitting)
}
