% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_cv.R
\name{edge_cv_sbm}
\alias{edge_cv_sbm}
\title{Edge CV for SBM (cross validation)}
\usage{
edge_cv_sbm(dat, k_vec, nfolds = 5, tol = 1e-06, verbose = T)
}
\arguments{
\item{dat}{\code{n} by \code{n} symmetric adjacency matrix}

\item{k_vec}{vector of positive integers, one per model, of how many blocks in said model}

\item{nfolds}{positive integer}

\item{tol}{small positive number}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Edge CV for SBM (cross validation)
}
