% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_cv.R
\name{edge_cv_sbm_sample_split}
\alias{edge_cv_sbm_sample_split}
\title{Edge CV for SBM (sample splitting)}
\usage{
edge_cv_sbm_sample_split(dat, k_vec, test_prop, tol = 1e-06)
}
\arguments{
\item{dat}{\code{n} by \code{n} symmetric adjacency matrix}

\item{k_vec}{vector of positive integers, one per model, of how many blocks in said model}

\item{test_prop}{test proportion}

\item{tol}{small positive number}
}
\value{
list
}
\description{
Edge CV for SBM (sample splitting)
}
