% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_cv_tensor.R
\name{edge_cv_sbm_tensor}
\alias{edge_cv_sbm_tensor}
\title{Edge CV for tensors (cross validation)}
\usage{
edge_cv_sbm_tensor(
  dat,
  k_vec,
  trials = 5,
  test_prop = 0.1,
  tol = 1e-06,
  verbose = T
)
}
\arguments{
\item{dat}{\code{p} by \code{n} by \code{n} symmetric adjacency matrix}

\item{k_vec}{vector of positive integers, one per model, of how many blocks in said model}

\item{trials}{positive integer}

\item{test_prop}{number between 0 and 1}

\item{tol}{small positive number}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Edge CV for tensors (cross validation)
}
