% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sbm.R
\name{generate_sbm}
\alias{generate_sbm}
\title{Generate SBM}
\usage{
generate_sbm(b_mat, cluster_idx, rho = 1)
}
\arguments{
\item{b_mat}{connectivity matrix (symmetric), all values between \code{0} and \code{1}}

\item{cluster_idx}{vector of length equal to number of nodes, where each node is 
assigned to a positive integer from \code{1} to \code{max(cluster_idx)}}

\item{rho}{sparsity index, between \code{0} and \code{1}}
}
\value{
symmetric \code{0}-\code{1} matrix
}
\description{
Generate SBM
}
