% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sbm.R
\name{generate_tensor}
\alias{generate_tensor}
\title{Generate tensor SBM}
\usage{
generate_tensor(b_tensor, cluster_idx, rho)
}
\arguments{
\item{b_tensor}{connectivity tensor as an \code{array}, organized so the dimension is \code{p} by \code{n} by \code{n},
representing \code{p} symmetric matrices all values between \code{0} and \code{1}}

\item{cluster_idx}{vector of length equal to the number of nodes \code{n}, where each node is 
assigned to a positive integer from \code{1} to \code{max(cluster_idx)}}

\item{rho}{sparsity index, between \code{0} and \code{1}}
}
\value{
a \code{array}
}
\description{
Generate tensor SBM
}
